remove depenency on bangle.dev lib
