+++ components/common/CharmEditor/components/@bangle.dev/tooltip/popper.ts
@@ -1,0 +1,6 @@
+export { default as arrow } from '@popperjs/core/lib/modifiers/arrow';
+export { default as flip } from '@popperjs/core/lib/modifiers/flip';
+export { default as offset } from '@popperjs/core/lib/modifiers/offset';
+export { default as popperOffsets } from '@popperjs/core/lib/modifiers/popperOffsets';
+export { default as preventOverflow } from '@popperjs/core/lib/modifiers/preventOverflow';
+export { createPopper } from '@popperjs/core/lib/popper-lite';
