+++ components/common/CharmEditor/components/emojiSuggest/emojiSuggest.plugins.ts
@@ -4,7 +4,7 @@ import { Plugin, PluginKey } from '@bangle.dev/pm';
 import type { SpecRegistry } from 'components/common/CharmEditor/components/@bangle.dev/core/specRegistry';
 
 import { createTooltipDOM } from '../@bangle.dev/tooltip';
-import * as suggestTooltip from '../@bangle.dev/tooltip/suggest-tooltip';
+import * as suggestTooltip from '../@bangle.dev/tooltip/suggestTooltip';
 
 import { markName } from './emojiSuggest.constants';
