+++ components/common/CharmEditor/components/listItem/listItem.component.ts
@@ -1,10 +1,9 @@
 import type { Command, EditorState, Node, Schema } from '@bangle.dev/pm';
-import { chainCommands, keymap } from '@bangle.dev/pm';
 import type { MoveDirection } from '@bangle.dev/pm-commands';
 import { copyEmptyCommand, cutEmptyCommand, moveNode, parentHasDirectParentOfType } from '@bangle.dev/pm-commands';
 import { browser, domSerializationHelpers, filter, insertEmpty, createObject } from '@bangle.dev/utils';
-import type Token from 'markdown-it/lib/token';
-import type { MarkdownSerializerState } from 'prosemirror-markdown';
+import { chainCommands } from 'prosemirror-commands';
+import { keymap } from 'prosemirror-keymap';
 
 import type { RawPlugins } from 'components/common/CharmEditor/components/@bangle.dev/core/plugin-loader';
 import type { RawSpecs } from 'components/common/CharmEditor/components/@bangle.dev/core/specRegistry';
