+++ components/common/CharmEditor/components/listItemNew/listItemPlugins.ts
@@ -1,6 +1,6 @@
-import { wrappingInputRule } from '@bangle.dev/pm';
 import { parentHasDirectParentOfType } from '@bangle.dev/pm-commands';
 import { filter } from '@bangle.dev/utils';
+import { wrappingInputRule } from 'prosemirror-inputrules';
 import { keymap } from 'prosemirror-keymap';
 import type { EditorState } from 'prosemirror-state';
 import { Plugin } from 'prosemirror-state';
