+++ components/common/CharmEditor/components/orderedList.ts
@@ -1,8 +1,9 @@
 import type { Command, EditorState, Schema } from '@bangle.dev/pm';
-import { keymap, wrappingInputRule } from '@bangle.dev/pm';
 import { parentHasDirectParentOfType } from '@bangle.dev/pm-commands';
 import { createObject } from '@bangle.dev/utils';
 import type Token from 'markdown-it/lib/token';
+import { wrappingInputRule } from 'prosemirror-inputrules';
+import { keymap } from 'prosemirror-keymap';
 
 import type { RawPlugins } from 'components/common/CharmEditor/components/@bangle.dev/core/plugin-loader';
 import type { RawSpecs } from 'components/common/CharmEditor/components/@bangle.dev/core/specRegistry';
