+++ lib/prosemirror/highlightMarkedElement.ts
@@ -1,7 +1,7 @@
 import type { EditorView, PluginKey } from '@bangle.dev/pm';
 import { TextSelection } from '@bangle.dev/pm';
 
-import { renderSuggestionsTooltip } from 'components/common/CharmEditor/components/@bangle.dev/tooltip/suggest-tooltip';
+import { renderSuggestionsTooltip } from 'components/common/CharmEditor/components/@bangle.dev/tooltip/suggestTooltip';
 import { highlightDomElement } from 'lib/utils/browser';
 
 export function highlightMarkedElement({
