+++ lib/prosemirror/interfaces.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-use-before-define */
 import type { PluginKey } from 'prosemirror-state';
 
-import type { PluginState as SuggestTooltipPluginState } from 'components/common/CharmEditor/components/@bangle.dev/tooltip/suggest-tooltip';
+import type { PluginState as SuggestTooltipPluginState } from 'components/common/CharmEditor/components/@bangle.dev/tooltip/suggestTooltip';
 
 export interface SuggestPluginState {
   tooltipContentDOM: HTMLElement;
