+++ package.json
@@ -90,7 +90,6 @@
     "@bangle.dev/base-components": "^0.31.6",
     "@bangle.dev/pm": "^0.31.6",
     "@bangle.dev/pm-commands": "^0.31.6",
-    "@bangle.dev/tooltip": "^0.31.6",
     "@bangle.dev/utils": "^0.31.6",
     "@beam-australia/react-env": "^3.1.1",
     "@ceramicnetwork/http-client": "^3.2.0",
@@ -136,6 +135,7 @@
     "@mux/upchunk": "^3.0.0",
     "@noble/ed25519": "^2.0.0",
     "@notionhq/client": "^2.2.2",
+    "@popperjs/core": "^2.11.8",
     "@react-dnd/shallowequal": "^4.0.0",
     "@react-email/column": "0.0.7",
     "@react-email/container": "0.0.8",
@@ -222,6 +222,7 @@
     "prosemirror-dropcursor": "^1.6.1",
     "prosemirror-example-setup": "^1.2.1",
     "prosemirror-gapcursor": "^1.3.1",
+    "prosemirror-history": "^1.4.0",
     "prosemirror-inputrules": "^1.2.1",
     "prosemirror-keymap": "^1.2.0",
     "prosemirror-markdown": "^1.10.1",
