+++ theme/@bangle.dev/styles.scss
@@ -624,3 +624,43 @@ ol li {
 li .insertion > .charm-tab {
   display: none;
 }
+
+/** Copied from tooltip **/
+.bangle-tooltip {
+  display: none;
+  z-index: 30;
+}
+
+.bangle-tooltip[data-show] {
+  display: block;
+}
+
+.bangle-tooltip-arrow,
+.bangle-tooltip-arrow::before {
+  position: absolute;
+  width: 8px;
+  height: 8px;
+  z-index: -1;
+}
+
+.bangle-tooltip-arrow::before {
+  content: '';
+  transform: rotate(45deg);
+  background: #333;
+}
+
+.bangle-tooltip[data-popper-placement^='top'] > .bangle-tooltip-arrow {
+  bottom: -4px;
+}
+
+.bangle-tooltip[data-popper-placement^='bottom'] > .bangle-tooltip-arrow {
+  top: -4px;
+}
+
+.bangle-tooltip[data-popper-placement^='left'] > .bangle-tooltip-arrow {
+  right: -4px;
+}
+
+.bangle-tooltip[data-popper-placement^='right'] > .bangle-tooltip-arrow {
+  left: -4px;
+}
